refactor: PingPong 객체 제거
