+++ src/main/kotlin/com/nexters/bottles/bottle/domain/Bottle.kt
@@ -1,5 +1,6 @@
 package com.nexters.bottles.bottle.domain
 
+import com.nexters.bottles.bottle.domain.enum.PingPongStatus
 import com.nexters.bottles.global.BaseEntity
 import com.nexters.bottles.user.domain.User
 import java.time.LocalDateTime
@@ -27,6 +28,13 @@ class Bottle(
     val sourceUser: User,
 
     @Column
-    val expiredAt: LocalDateTime = LocalDateTime.now().plusDays(1)
+    val expiredAt: LocalDateTime = LocalDateTime.now().plusDays(1),
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "stopped_user_id")
+    var stoppedUser: User? = null,
+
+    @Column
+    val pingPongStatus: PingPongStatus = PingPongStatus.NONE,
 ) : BaseEntity() {
 }
