+++ src/main/kotlin/com/nexters/bottles/bottle/domain/enum/PingPongStatus.kt
@@ -1,6 +1,7 @@
-package com.nexters.bottles.pingpong.domain.enum
+package com.nexters.bottles.bottle.domain.enum
 
 enum class PingPongStatus {
+    NONE,
     ACTIVE,
     STOPPED,
     DONE,
