+++ src/main/kotlin/com/nexters/bottles/bottle/facade/BottleFacade.kt
@@ -3,13 +3,11 @@ package com.nexters.bottles.bottle.facade
 import com.nexters.bottles.bottle.facade.dto.BottleDto
 import com.nexters.bottles.bottle.facade.dto.BottleListResponseDto
 import com.nexters.bottles.bottle.service.BottleService
-import com.nexters.bottles.pingpong.service.PingPongService
 import org.springframework.stereotype.Component
 
 @Component
 class BottleFacade(
-    private val bottleService: BottleService,
-    private val pingPongService: PingPongService
+    private val bottleService: BottleService
 ) {
 
     fun getBottles(): BottleListResponseDto {
@@ -30,7 +28,6 @@ class BottleFacade(
     }
 
     fun acceptBottle(bottleId: Long) {
-        val users = bottleService.validateBottleUsers(bottleId)
-        pingPongService.startPingPong(users[0], users[1])
+        bottleService.acceptBottle(bottleId)
     }
 }
