+++ src/main/kotlin/com/nexters/bottles/bottle/repository/LetterRepository.kt
@@ -1,6 +1,6 @@
-package com.nexters.bottles.pingpong.repository
+package com.nexters.bottles.bottle.repository
 
-import com.nexters.bottles.pingpong.domain.Letter
+import com.nexters.bottles.bottle.domain.Letter
 import org.springframework.data.jpa.repository.JpaRepository
 
 interface LetterRepository : JpaRepository<Letter, Long> {
