+++ src/main/kotlin/com/nexters/bottles/bottle/repository/QuestionRepository.kt
@@ -1,6 +1,6 @@
-package com.nexters.bottles.pingpong.repository
+package com.nexters.bottles.bottle.repository
 
-import com.nexters.bottles.pingpong.domain.Question
+import com.nexters.bottles.bottle.domain.Question
 import org.springframework.data.jpa.repository.JpaRepository
 
 interface QuestionRepository : JpaRepository<Question, Long> {
