+++ src/main/kotlin/com/nexters/bottles/bottle/service/BottleService.kt
@@ -1,7 +1,11 @@
 package com.nexters.bottles.bottle.service
 
 import com.nexters.bottles.bottle.domain.Bottle
+import com.nexters.bottles.bottle.domain.Letter
+import com.nexters.bottles.bottle.domain.LetterQuestionAndAnswer
 import com.nexters.bottles.bottle.repository.BottleRepository
+import com.nexters.bottles.bottle.repository.LetterRepository
+import com.nexters.bottles.bottle.repository.QuestionRepository
 import com.nexters.bottles.user.domain.User
 import com.nexters.bottles.user.repository.UserRepository
 import org.springframework.data.repository.findByIdOrNull
@@ -12,7 +16,9 @@ import java.time.LocalDateTime
 @Service
 class BottleService(
     private val bottleRepository: BottleRepository,
-    private val userRepository: UserRepository
+    private val userRepository: UserRepository,
+    private val letterRepository: LetterRepository,
+    private val questionRepository: QuestionRepository
 ) {
 
     @Transactional(readOnly = true)
@@ -23,8 +29,8 @@ class BottleService(
         return bottleRepository.findByTargetUserAndNotExpired(user, LocalDateTime.now())
     }
 
-    @Transactional(readOnly = true)
-    fun validateBottleUsers(bottleId: Long): List<User> {
+    @Transactional
+    fun acceptBottle(bottleId: Long) {
         val bottle = bottleRepository.findByIdAndNotExpired(bottleId, LocalDateTime.now())
             ?: throw IllegalArgumentException("이미 떠내려간 보틀이에요")
 
@@ -33,6 +39,24 @@ class BottleService(
         val sourceUser = userRepository.findByIdOrNull(bottle.sourceUser.id)
             ?: throw IllegalArgumentException("탈퇴한 회원이에요")
 
-        return listOf(targetUser, sourceUser)
+        val letters = findRandomQuestions()
+        saveLetter(bottle, targetUser, letters)
+        saveLetter(bottle, sourceUser, letters)
+    }
+
+    private fun findRandomQuestions() = questionRepository.findAll()
+        .shuffled()
+        .take(3)
+        .map {
+            LetterQuestionAndAnswer(question = it.question)
+        }
+
+    private fun saveLetter(
+        bottle: Bottle,
+        user: User,
+        letters: List<LetterQuestionAndAnswer>
+    ) {
+        val letter = Letter(bottle = bottle, user = user, letters = letters)
+        letterRepository.save(letter)
     }
 }
