+++ src/main/resources/sql/table_query.sql
@@ -21,35 +21,26 @@ CREATE TABLE user_profile
 
 CREATE TABLE bottle
 (
-    id             BIGINT AUTO_INCREMENT PRIMARY KEY,
-    target_user_id BIGINT                             NOT NULL,
-    source_user_id BIGINT                             NOT NULL,
-    expired_at     DATETIME                           NOT NULL,
-    created_at     DATETIME DEFAULT CURRENT_TIMESTAMP NOT NULL,
-    updated_at     DATETIME DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP NOT NULL
-)
-
-CREATE TABLE ping_pong
-(
-    id              BIGINT AUTO_INCREMENT PRIMARY KEY,
-    user_a_id       BIGINT                                NOT NULL,
-    user_b_id       BIGINT                                NOT NULL,
-    stopped_user_id BIGINT,
-    status          VARCHAR(20) DEFAULT 'ACTIVE'          NOT NULL,
-    created_at      DATETIME    DEFAULT CURRENT_TIMESTAMP NOT NULL,
-    updated_at      DATETIME    DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP NOT NULL
+    id               BIGINT AUTO_INCREMENT PRIMARY KEY,
+    target_user_id   BIGINT                                NOT NULL,
+    source_user_id   BIGINT                                NOT NULL,
+    expired_at       DATETIME                              NOT NULL,
+    stopped_user_id  BIGINT,
+    ping_pong_status VARCHAR(20) DEFAULT 'NONE'            NOT NULL,
+    created_at       DATETIME    DEFAULT CURRENT_TIMESTAMP NOT NULL,
+    updated_at       DATETIME    DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP NOT NULL
 )
 
 CREATE TABLE letter
 (
-    id           BIGINT AUTO_INCREMENT PRIMARY KEY,
-    ping_pong_id BIGINT                             NOT NULL,
-    user_id      BIGINT                             NOT NULL,
-    letters      JSON                               NOT NULL,
-    image        TEXT,
-    is_read      BOOLEAN  DEFAULT FALSE             NOT NULL,
-    created_at   DATETIME DEFAULT CURRENT_TIMESTAMP NOT NULL,
-    updated_at   DATETIME DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP NOT NULL
+    id         BIGINT AUTO_INCREMENT PRIMARY KEY,
+    bottle_id  BIGINT                             NOT NULL,
+    user_id    BIGINT                             NOT NULL,
+    letters    JSON                               NOT NULL,
+    image      TEXT,
+    is_read    BOOLEAN  DEFAULT FALSE             NOT NULL,
+    created_at DATETIME DEFAULT CURRENT_TIMESTAMP NOT NULL,
+    updated_at DATETIME DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP NOT NULL
 )
 
 CREATE TABLE question
